Move SpeedSlider to separate file
