+++ lib/home.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:marklin_bluetooth/controller.dart';
+import 'package:marklin_bluetooth/controller_screen/controller.dart';
 import 'package:marklin_bluetooth/current_race_screen/current_race_screen.dart';
 import 'package:marklin_bluetooth/race_browser_screen/race_browser.dart';
